rental return to base calculate

+++ src/CoreGoDelivery.Domain/Consts/RentalServiceConst.cs
@@ -3,9 +3,9 @@
 public static class RentalServiceConst
 {
     public const string CURRENCY_BRL = "R$";
-    public const int PENALTY_VALUE_PER_DAY = 50;
+    public const double PENALTY_VALUE_PER_DAY = 50.0;
     public const int MINIMAL_DAYS_PLAN = 7;
-    public const int MINIMAL_FEE_PERCENTAGE = 20;
-    public const int DEFAULT_FEE_PERCENTAGE = 40;
+    public const double MINIMAL_FEE_PERCENTAGE = 20.0;
+    public const double DEFAULT_FEE_PERCENTAGE = 40.0;
     public const string MESSAGE_RETURNED_TO_BASE_SUCCESS = "return to base date accepted with success";
 }
